Evolutionary Distinctiveness and Global Endangerment

This function calculates EDGE by combining evolutionary distinctiveness (ED; i.e., phylogenetic isolation of a species) with global endangerment (GE) status as defined by the International Union for Conservation of Nature (IUCN).

EDGE(x, phy, Redlist = "Redlist", species = "species", ...)

Arguments

x
a data.frame

phy
a phylogenetic tree (object of class phylo).

Redlist
column in the data frame with the IUCN ranks: LC, NT, VU, EN, CR, and EX.

species
data frame column specifying the taxon

...
Further arguments passed to or from other methods.

Value

Returns a dataframe of EDGE scores

Details

EDGE is calculated as: $$log(1+ED) + GE*log(2)$$ where ED represents the evolutionary distinctiveness score of each species (function evol_distinct), i.e. the degree of phylogenetic isolation, and combining it with GE, global endangerment from IUCN conservation threat categories. GE is calculated as the expected probability of extinction over 100 years of each taxon in the phylogeny (Redding & Mooers, 2006), scaled as follows: least concern = 0.001, near threatened and conservation dependent = 0.01, vulnerable = 0.1, endangered = 0.67, and critically endangered = 0.999.

References

Redding, D.W., & Mooers, A.Ø. (2006) Incorporating evolutionary measures into conservation prioritization. Conservation Biology 20: 1670–1678.

Isaac, N.J., Turvey, S.T., Collen, B., Waterman, C. & Baillie, J.E. (2007) Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2: e296.

Author

Barnabas H. Daru

Examples

data(africa)
y <- EDGE(africa$IUCN, africa$phylo, Redlist = "IUCN", species="Species")