Fits Grade of membership models for phyloregions

This function generates grade of membership, “admixture”, “topic” or “Latent Dirichlet Allocation” models, for clustering regions based on attributes.

admixture(
  counts,
  K,
  shape = NULL,
  initopics = NULL,
  tol = 0.1,
  bf = TRUE,
  kill = 2,
  ord = TRUE,
  verb = 1,
  ...
)

Arguments

counts
A matrix of multinomial response counts in ncol(counts) phrases/categories for nrow(counts) documents/observations. It can also be a simple matrix.

K
The number of latent topics. If length(K)>1, topics will find the Bayes factor (vs a null single topic model) for each element and return parameter estimates for the highest probability K.

shape
Optional argument to specify the Dirichlet prior concentration parameter as shape for topic-phrase probabilities. Defaults to 1/(K*ncol(counts)). For fixed single K, this can also be a ncol(counts) by K matrix of unique shapes for each topic element.

initopics
Optional start-location for [θ_1 ... θ_K], the topic-phrase probabilities. Dimensions must accord with the smallest element of K. If NULL, the initial estimates are built by incrementally adding topics.

tol
An indicator for whether or not to calculate the Bayes factor for univariate K. If length(K)>1, this is ignored and Bayes factors are always calculated.

bf
An indicator for whether or not to calculate the Bayes factor for univariate K. If length(K)>1, this is ignored and Bayes factors are always calculated.

kill
For choosing from multiple K numbers of topics (evaluated in increasing order), the search will stop after kill consecutive drops in the corresponding Bayes factor. Specify kill=0 if you want Bayes factors for all elements of K.

ord
If TRUE, the returned topics (columns of theta) will be ordered by decreasing usage (i.e., by decreasing colSums(omega)).

verb
A switch for controlling printed output. verb > 0 will print something, with the level of detail increasing with verb.

...
Further arguments passed to or from other methods.

Value

An topics object list with entries

K The number of latent topics estimated. If input length(K)>1, on output this is a single value corresponding to the model with the highest Bayes factor.

theta The ncolcounts by K matrix of estimated topic-phrase probabilities.

omega The nrowcounts by K matrix of estimated document-topic weights.

BF The log Bayes factor for each number of topics in the input K, against a null single topic model.

D Residual dispersion: for each element of K, estimated dispersion parameter (which should be near one for the multinomial), degrees of freedom, and p-value for a test of whether the true dispersion is >1.

X The input count matrix, in dgTMatrix format.