Bin values

choropleth discretizes the values of a quantity for mapping.

choropleth(x, k = 10, breaks = "quantile", min = NULL, max = NULL)

Arguments

x

Vector of values to discretize.

k

Numeric, the desired number of bins to discretize.

breaks

one of “equal”, “pretty”, “jenks”, “quantile” or numeric vector with the actual breaks by specifying the minimum (min) and maximum (max) bounds.

min

the minima of the lowest bound of the break.

max

the maxima of the upper bound of the break

Value

a vector with the discretized values.

See also

Author

Barnabas H. Daru darunabas@gmail.com