Species' evolutionary distinctiveness

Calculates evolutionary distinctiveness measures for a suite of species by: a) equal splits (Redding and Mooers 2006) b) fair proportions (Isaac et al., 2007). This a new implementation of the picante function evol.distinct however allowing multifurcations and can be orders of magnitude faster.

evol_distinct(
  tree,
  type = c("equal.splits", "fair.proportion"),
  scale = FALSE,
  use.branch.lengths = TRUE,
  ...
)

Arguments

tree
an object of class phylo.

type
a) equal splits (Redding and Mooers 2006) or b) fair proportions (Isaac et al., 2007)

scale
The scale option refers to whether or not the phylogeny should be scaled to a depth of 1 or, in the case of an ultrametric tree, scaled such that branch lengths are relative.

use.branch.lengths
If use.branch.lengths=FALSE, then all branch lengths are changed to 1.

...
Further arguments passed to or from other methods.

Value

a named vector with species scores.

References

Redding, D.W. and Mooers, A.O. (2006). Incorporating evolutionary measures into conservation prioritisation. Conservation Biology, 20, 1670--1678.

Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C. and Baillie, J.E.M. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE, 2, e296.

See also

Author

Klaus Schliep

Examples

tree <- ape::rcoal(10)
evol_distinct(tree)

#>        t6        t8        t5        t1        t7        t3        t2       t10 
#> 2.3213056 1.7169957 0.5710788 0.5710788 0.3468188 0.3468188 0.1715147 0.1715147 
#>        t4        t9 
#> 0.1740371 0.1740371 

evol_distinct(tree, type = "fair.proportion")

#>        t6        t8        t5        t1        t7        t3        t2       t10 
#> 2.3213056 1.2469769 0.4575321 0.4575321 0.3886116 0.3886116 0.3248963 0.3248963 
#>        t4        t9 
#> 0.3274188 0.3274188