Create a fishnet of regular grids

The fishnet function creates a regular grid of locations covering the study area at various grain sizes.

fishnet(mask, res = 0.5)

Arguments

mask

a vector polygon covering the boundary of the survey region.

res

the grain size of the grid cells in decimal degrees (default).

Value

A spatial vector polygon object of equal area grid cells covering the defined area.

References

Phillips, S.J., Anderson, R.P. & Schapire, R.E. (2006) Maximum entropy modeling of species geographic distributions. Ecological Modelling 190: 231-259.

Examples

d <- terra::vect(system.file("ex/nigeria.json", package="phyloregion"))
f <- fishnet(d, res = 0.75)