Calculate evolutionary distinctiveness of phyloregions

This function estimates evolutionary distinctiveness of each phyloregion by computing the mean value of phylogenetic beta diversity between a focal phyloregion and all other phyloregions in the study area.

phyloregion(x, k = 10, method = "average", shp = NULL, ...)

infomap(x, shp = NULL, ...)

Arguments

x
A distance matrix

k
The desired number of phyloregions, often as determined by optimal_phyloregion.

method
the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of “ward.D”, “ward.D2”, “single”, “complete”, “average” (= UPGMA), “mcquitty” (= WPGMA), “median” (= WPGMC) or “centroid” (= UPGMC).

shp
a polygon shapefile of grid cells.

...
Further arguments passed to or from other methods.

Value

An object of class phyloregion containing

a data frame membership with columns grids and cluster

k the number of clusters and additionally there can be an shape file and other bjects. This representation may still change.

References

Daru, B.H., Van der Bank, M., Maurin, O., Yessoufou, K., Schaefer, H., Slingsby, J.A. & Davies, T.J. (2016) A novel phylogenetic regionalization of the phytogeographic zones of southern Africa reveals their hidden evolutionary affinities. Journal of Biogeography 43: 155-166.

Daru, B.H., Elliott, T.L., Park, D.S. & Davies, T.J. (2017) Understanding the processes underpinning patterns of phylogenetic regionalization. Trends in Ecology and Evolution 32: 845-860.

Daru, B.H., Holt, B.G., Lessard, J.P., Yessoufou, K. & Davies, T.J. (2017) Phylogenetic regionalization of marine plants reveals close evolutionary affinities among disjunct temperate assemblages. Biological Conservation 213: 351-356.

See also

Author

Barnabas H. Daru darunabas@gmail.com

Examples

library(ape)
tree <- read.tree(text = "((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- sparseMatrix(c(1,3,4,1,4,5,1,2,3,4,5,6,3,4,6),
  c(1,1,1,2,2,2,3,3,3,3,3,3,4,4,4),x=1,
  dimnames = list(paste0("g", 1:6), tree$tip.label))
pbc <- phylobeta(com, tree)
phyloregion(pbc[[1]], k = 3)

#> $membership
#>   grids cluster
#> 1    g1       1
#> 2    g2       1
#> 3    g3       2
#> 4    g4       1
#> 5    g5       1
#> 6    g6       3
#> 
#> $k
#> [1] 3
#> 
#> $evol_distinct
#>   cluster         ED
#> 1       1 0.13958333
#> 2       2 0.05625000
#> 3       3 0.08333333
#> 
#> $region.dist
#>           1         2
#> 2 0.1125000          
#> 3 0.1666667 0.0000000
#> 
#> $region.df
#>   grids cluster
#> 1    g1       1
#> 2    g2       1
#> 3    g3       2
#> 4    g4       1
#> 5    g5       1
#> 6    g6       3
#> 
#> attr(,"class")
#> [1] "phyloregion"