Biogeographic regionalization and macroecology

This document describes the phyloregion package for the R software. phyloregion is a computational infrastructure for biogeographic regionalization (the classification of geographical areas in terms of their biotas) and spatial conservation in the R scientific computing environment. Previous analyses of biogeographical regionalization were either focused on smaller datasets or slower particularly when the number of species or geographic scale is very large. With macroecological datasets of ever increasing size and complexity, phyloregion offers the possibility of handling and executing large scale biogeographic regionalization efficiently and with extreme speed. It also allows fast and efficient for analysis of more standard conservation measures such as phylogenetic diversity, phylogenetic endemism, evolutionary distinctiveness and global endangerment. phyloregion can run on any operating system (Mac, Linux, Windows or even high performance computing cluster) with R 3.6.0 (or higher) installed.

How to cite phyloregion

The original implementation of phyloregion is described in:

Daru B.H., Karunarathne, P. & Schliep, K. (2020) phyloregion: R package for biogeographic regionalization and macroecology. Methods in Ecology and Evolution 11, 1483-1491.

It is based on the method described in:

Daru, B.H., Farooq, H., Antonelli, A. & Faurby, S. (2020) Endemism patterns are scale dependent. Nature Communications 11, 2115.

The original conceptual is described in:

Daru, B.H., Elliott, T.L., Park, D.S. & Davies, T.J. (2017) Understanding the processes underpinning patterns of phylogenetic regionalization. Trends in Ecology and Evolution 32: 845-860.

Feedback

If you have any questions, suggestions or issues regarding the package, please add them to GitHub issues

Installation

phyloregion is an open-source and free package hosted on GitHub. You will need to install the devtools package. In R, type:

if (!requireNamespace("devtools", quietly = TRUE)) install.packages("devtools")

Then:

devtools::install_github("darunabas/phyloregion")

Load the phyloregion package:

library(phyloregion)

Acknowledgments

Barnabas Daru thanks Texas A&M University-Corpus Christi for financial and logistic support.

Author

Barnabas H. Daru, Piyal Karunarathne, Klaus Schliep