Convert raw input distribution data to community

The functions points2comm, polys2comm, raster2comm provide convenient interfaces to convert raw distribution data often available as point records, extent-of-occurrence polygons and raster layers, respectively, to a community composition data frame at varying spatial grains and extents for downstream analyses.

raster2comm(files)

polys2comm(dat, res = 1, shp.grids = NULL, species = "species", ...)

points2comm(
  dat,
  mask = NULL,
  res = 1,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  shp.grids = NULL,
  ...
)

Arguments

files
list of raster layer objects with the same spatial extent and resolution.

dat
layers of merged maps corresponding to species ranges for polys2comm; or point occurrence data frame for points2comm, with at least three columns:

Column 1: species (listing the taxon names)

Column 2: decimallongitude (corresponding to decimal longitude)

Column 3: decimallatitude (corresponding to decimal latitude)

res
the grain size of the grid cells in decimal degrees (default).

shp.grids
if specified, the polygon shapefile of grid cells with a column labeled “grids”.

species
a character string. The column with the species or taxon name. Default = “species”.

...
Further arguments passed to or from other methods.

mask
Only applicable to points2comm. If supplied, a polygon shapefile covering the boundary of the survey region.

lon
character with the column name of the longitude.

lat
character with the column name of the latitude.

Value

comm_dat: community data frame

poly_shp: shapefile of grid cells with the values per cell.

See also

Examples

# \donttest{
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
raster2comm(files)

#> $comm_dat
#> 207 x 16 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 16 column names ‘Chytranthus_gilletii’, ‘Commelina_ramulosa’, ‘Cymbopogon_caesius’ ... ]]

#>                                     
#> v100 . . . 1 . . . . . . . . . . . .
#> v101 . . . 1 . . . . . . . . . . . .
#> v102 . . . 1 . . . . . . . . . . . .
#> v103 . . . 1 . . . . . . . . . . . .
#> v104 . . . 1 . . . . . . . . . . . .
#> v105 . . . 1 . . . . . . . . . . . .
#> v106 . . . 1 . . . . . . . . . . . .
#> v107 . . . 1 . . . . . . . . . . . .
#> v123 . 1 . . . . . . . . . . . . . .
#> v124 . 1 . . . . . . . . . . . . . .
#> v125 . 1 . . . . . . . . . . . . . .
#> v148 . 1 . . . 1 . . . . . . . . . .
#> v149 . . . . . 1 . . . . . . . . . .
#> v150 . . . . . 1 . . . . . . . . . .
#> v151 . . . . . 1 . . . . . . . . . .
#> v152 . . . 1 . 1 . . . . . . . . . .
#> v153 1 . . 1 . 1 . . . . . . . . . .
#> v154 1 . . . . . . . . . . . . . . .
#> v155 1 . . . . . . . . . . . . . . .
#> v156 1 . . . . . . . . . . . . . . .
#> v158 . . . . . . . . . . . . . . 1 .
#> v171 . 1 . 1 . 1 . 1 . . . . . . . .
#> v172 . 1 . . . 1 . 1 . . . . . . . .
#> v173 . . . . . 1 . 1 . . . . . . . .
#> v174 . . . 1 . 1 . 1 . . . . . . . .
#> v175 . . . 1 . 1 . 1 . . . . . . . .
#> v176 1 . . 1 . 1 . . . . . . . . . .
#> v177 1 . . 1 . 1 . . . . . . . . 1 .
#> v178 1 . . 1 . 1 . . . . . . . . 1 .
#> v179 1 . . 1 . . . . . . . . . . 1 .
#> v180 1 . . 1 . . . . . . . . . . . .
#> v181 1 . . . . . . . . . . . . . 1 .
#> v182 1 . . . . . . . . . . . . . 1 .
#> v195 . 1 . 1 . 1 . 1 . . . . . . 1 .
#> v196 . 1 . 1 . 1 . 1 . . . . . . 1 .
#> v197 . . . 1 . 1 . 1 . . . . . . 1 .
#> v198 . . . 1 . 1 . 1 . . . . . . 1 .
#> v199 1 . . 1 . 1 . 1 . . . . . . 1 .
#> v200 1 . . 1 . . . . 1 . . . . . 1 .
#> v201 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v202 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v203 1 . . 1 . . . . 1 . . . . . 1 .
#> v204 1 . . 1 . . . . . . . . . . 1 .
#> v205 1 . . 1 . . . . . . . . . . 1 .
#> v206 1 . . . . . . . . . . . . . 1 .
#> v207 1 . . . . . . . . . . . . . 1 .
#> v208 . . . 1 . . . . . . . . . . . .
#> v209 . . . 1 . . . . . . . . . . . .
#> v218 . . 1 1 . 1 . 1 1 . . . . . 1 .
#> v219 1 . 1 1 . 1 . 1 1 . . . . . 1 .
#> v220 . . 1 1 . 1 . 1 1 . . . . . 1 .
#> v221 1 1 1 1 . 1 . 1 1 . . . . . 1 .
#> v222 1 . 1 1 . 1 . 1 1 . . 1 . . 1 .
#> v223 1 1 1 1 . 1 . 1 1 . . 1 . . 1 .
#> v224 1 1 1 1 . . . . 1 . . . . . 1 .
#> v225 1 . . 1 . . . . 1 . . . . . 1 .
#> v226 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v227 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v228 1 . . 1 . . . . 1 . . . 1 . . .
#> v229 1 . . 1 . . . . . . . . . . 1 .
#> v230 1 . . 1 . . . . . . . . . . 1 .
#> v231 1 . . 1 . . . . . . . . . . 1 .
#> v232 . . . . . . . . . . . . . . . 1
#> v233 . . . . . . . . . . . . . . . 1
#> v234 . . . . . . . . . . . . . . . 1
#> v241 . . 1 1 1 1 . 1 . . . 1 . . 1 .
#> v242 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v243 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v244 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v245 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v246 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v247 1 1 1 1 . 1 . . 1 . . . . . 1 .
#> v248 1 1 1 1 . 1 . . 1 . . . . . 1 .
#> v249 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v250 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v251 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v252 1 . . 1 . . . . 1 . . . . . 1 .
#> v253 1 . . 1 . . . . 1 . . . . . 1 .
#> v254 1 . . 1 . . . . 1 . . . . . 1 1
#> v255 1 . . 1 . . . . 1 . . . . . 1 1
#> v256 . . . . . . . . 1 . . . . . 1 1
#> v257 . . . . . . . . . . . . . . . 1
#> v258 . . . . . . . . . . . . . . . 1
#> v259 . . . . . . . . . . . . . . 1 .
#> v265 1 . 1 1 1 1 . 1 . . . 1 . . 1 .
#> v266 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v267 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v268 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v269 1 . 1 1 1 1 . 1 1 . . . . . 1 .
#> v270 1 . 1 1 1 1 . . 1 . . . . . 1 .
#> v271 1 . 1 1 1 1 . . 1 . . . . . 1 .
#> v272 1 . 1 1 1 1 . . 1 . . . . . 1 .
#> v273 1 1 1 1 1 1 . . 1 . . . . . 1 .
#> v274 1 . . 1 1 1 . . 1 . . . . . 1 .
#> v275 1 . . 1 1 1 . . 1 . . . . . 1 .
#> v276 1 . . 1 . 1 . . 1 . . . . . 1 .
#> v277 1 . . 1 . . . . 1 . . . . . 1 1
#> v278 1 . . 1 . . . . 1 . . . . . 1 1
#> v279 1 . . . . . . . 1 . . . . . 1 1
#> v28  . . . . . . 1 . . . . . . . . .
#> v280 1 . . . . . . . 1 . . . . . 1 1
#> v281 1 . . 1 . . . . 1 . . . . . 1 1
#> v282 1 . . 1 . . . . . . . . . . . 1
#> v283 . . . . . . . . . . . . . . 1 1
#> v289 1 . 1 1 1 1 . 1 . . . 1 . . 1 .
#> v29  . . . . . . 1 . . . . . . . . .
#> v290 1 . 1 1 1 1 . . . . 1 1 . . 1 .
#> v291 1 . 1 1 1 1 . . 1 . . 1 . . 1 .
#> v292 1 . 1 1 1 1 . . 1 . 1 1 . . 1 .
#> v293 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v294 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v295 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v296 1 . 1 1 1 1 . . 1 . . . . . 1 .
#> v297 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v298 1 . . 1 1 1 . . 1 . . . . . 1 .
#> v299 1 . . 1 1 1 . . 1 . . . . . 1 .
#> v30  . . . . . . 1 . . . . . . . . .
#> v300 1 . . 1 1 1 . . 1 . . . . . 1 .
#> v301 1 . . 1 . . . . 1 . . . . . 1 1
#> v302 1 . . 1 . . . . 1 . . . . . 1 1
#> v303 1 . . 1 . . . . 1 . . . . . 1 1
#> v304 1 . . 1 . . . . 1 . . . . . 1 1
#> v305 1 . . 1 . . . . 1 . . . . . 1 1
#> v306 1 . . 1 . . . . 1 . . . . . . 1
#> v307 1 . . 1 . . . . 1 . . . . . . 1
#> v31  . . . . . . 1 . . . . . . . . .
#> v314 1 . 1 1 1 1 . . . . 1 1 . . 1 .
#> v315 1 . 1 1 1 1 . . 1 . 1 1 . . 1 .
#> v316 1 . 1 1 1 1 . . 1 . 1 1 . . 1 .
#> v317 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v318 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v319 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v32  . . . . . . 1 . . . . . . . . .
#> v320 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v321 1 . 1 1 1 1 . . 1 . 1 . . . 1 .
#> v322 1 . . 1 1 1 . . 1 . 1 . . . 1 .
#> v323 1 . . 1 1 1 . . 1 . . . . . . .
#> v324 1 . . 1 1 1 . . 1 . . . . . . .
#> v325 1 . . 1 1 . . . 1 . . . . . . .
#> v326 1 . . 1 1 . . . 1 . . . . . . 1
#> v327 1 . . 1 1 . . . 1 . . . . . . 1
#> v328 1 . . 1 1 . . . 1 . . . . . . 1
#> v329 1 . . 1 1 . . . 1 . . . . . . 1
#> v330 1 . . 1 1 . . . 1 . . . . . . .
#> v331 1 . . 1 . . . . 1 . . . . . . .
#> v338 1 . 1 1 1 1 . . . . 1 1 1 . 1 .
#> v339 1 . 1 1 1 1 . . . . 1 1 1 . . .
#> v340 1 . 1 1 1 1 . . 1 . 1 . 1 . . .
#> v341 1 . 1 1 1 1 . . 1 . 1 . 1 . . .
#> v342 1 . . 1 1 1 . . 1 . 1 . 1 . . .
#> v343 1 . . 1 1 1 . . 1 . 1 . 1 . 1 .
#> v344 1 . . 1 1 1 . . 1 1 1 . 1 . 1 .
#> v345 1 . . 1 1 1 . . 1 1 1 . 1 . . .
#> v346 1 . . 1 1 1 . . 1 1 . . 1 . . .
#> v347 1 . . 1 1 1 . . 1 1 . . 1 . . .
#> v348 1 . . 1 1 . . . 1 1 . . 1 . . .
#> v349 1 . . 1 1 . . . 1 1 . . 1 . . .
#> v35  . . . . . . 1 . . . . . . . . .
#> v350 1 . . 1 1 . . . 1 1 . . 1 . . 1
#> v351 1 . . 1 1 . . . 1 1 . . 1 . . 1
#> v354 1 . . 1 1 . . . 1 . . . 1 . . .
#> v365 1 . . 1 1 . . . . . 1 . 1 . . .
#> v366 1 . . 1 1 . . . . . 1 . 1 . . .
#> v367 1 . . 1 1 . . . 1 1 1 . 1 . 1 .
#> v368 1 . . 1 1 . . . 1 1 1 . 1 . 1 .
#> v369 1 . . 1 1 . . . 1 1 1 . 1 . . .
#> v370 1 . . 1 1 . . . 1 1 1 . 1 . . .
#> v371 1 . . 1 1 . . . 1 1 1 . 1 . . .
#> v372 1 . . 1 1 . . . 1 1 . . 1 1 . .
#> v373 1 . . 1 1 . . . 1 1 . . 1 1 . .
#> v374 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v390 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v391 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v392 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v393 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v394 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v395 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v396 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v397 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v415 1 . . 1 1 . . . . 1 1 1 1 1 . .
#> v416 1 . . 1 1 . . . . 1 1 1 1 1 . .
#> v417 1 . . 1 1 . . . 1 1 1 1 1 1 . .
#> v418 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v419 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v420 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v421 1 . . 1 1 . . . 1 1 1 . 1 1 . .
#> v439 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v440 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v441 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v442 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v443 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v444 1 . . 1 1 . . . . 1 1 . 1 1 . .
#> v5   . . . . . . 1 . . . . . . . . .
#> v52  . . . . . . 1 . . . . . . . . .
#> v53  . . . . . . 1 . . . . . . . . .
#> v54  . . . . . . 1 . . . . . . . . .
#> v55  . . . . . . 1 . . . . . . . . .
#> v56  . . . . . . 1 . . . . . . . . .
#> v57  . . . . . . 1 . . . . . . . . .
#> v58  . . . . . . 1 . . . . . . . . .
#> v59  . . . . . . 1 . . . . . . . . .
#> v7   . . . . . . 1 . . . . . . . . .
#> v76  . . . 1 . . . . . . . . . . . .
#> v77  . . . 1 . . . . . . . . . . . .
#> v78  . . . 1 . . . . . . . . . . . .
#> v80  . . . 1 . . . . . . . . . . . .
#> v99  . 1 . . . . . . . . . . . . . .
#> 
#> $poly_shp
#> class       : SpatialPolygonsDataFrame 
#> features    : 456 
#> extent      : 2.5, 14.5, 4.5, 14  (xmin, xmax, ymin, ymax)
#> crs         : +proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0 
#> variables   : 4
#> names       : grids, richness,   lon,   lat 
#> min values  :    v1,        0,  2.75,  4.75 
#> max values  :   v99,        1, 14.25, 13.75 
#> 

# }

# \donttest{
s <- readRDS(system.file("nigeria/nigeria.rds", package="phyloregion"))
sp <- random_species(100, species=5, shp=s)
polys2comm(dat = sp, species = "species")

#> $comm_dat
#> 92 x 5 sparse Matrix of class "dgCMatrix"
#>      species1 species2 species3 species4 species5
#> v100        1        1        .        .        1
#> v101        1        .        .        .        1
#> v105        1        .        .        .        .
#> v106        1        .        .        .        1
#> v107        1        .        .        .        1
#> v108        1        1        .        .        1
#> v109        1        1        .        .        1
#> v110        1        1        .        .        1
#> v111        1        1        .        .        1
#> v112        1        .        .        .        1
#> v113        1        .        .        .        1
#> v120        1        .        .        .        .
#> v121        1        .        .        .        1
#> v122        1        .        .        .        1
#> v123        1        .        .        .        1
#> v124        1        .        .        .        1
#> v135        .        .        .        .        1
#> v16         1        .        .        .        1
#> v17         1        .        1        .        1
#> v18         1        .        1        .        1
#> v19         1        1        .        .        1
#> v20         1        1        .        .        1
#> v21         1        1        .        1        1
#> v22         1        1        .        1        1
#> v23         1        1        .        1        1
#> v24         1        1        .        1        1
#> v25         1        1        .        .        1
#> v28         1        .        .        .        .
#> v29         1        1        .        .        1
#> v3          .        .        .        .        1
#> v30         1        1        1        .        1
#> v31         1        1        1        1        1
#> v32         1        1        .        1        1
#> v33         1        1        .        1        1
#> v34         1        1        .        1        1
#> v35         1        1        .        1        1
#> v36         1        1        .        1        1
#> v37         1        1        .        1        1
#> v38         1        1        .        .        1
#> v39         1        .        .        .        1
#> v41         1        .        .        .        .
#> v42         1        1        .        .        1
#> v43         1        1        1        1        1
#> v44         1        1        1        1        1
#> v45         1        1        .        1        1
#> v46         1        1        .        1        1
#> v47         1        1        .        1        1
#> v48         1        1        .        1        1
#> v49         1        1        .        1        1
#> v50         1        1        .        1        1
#> v51         1        1        .        .        1
#> v52         .        1        .        .        1
#> v54         1        .        .        .        1
#> v55         1        1        .        1        1
#> v56         1        1        .        1        1
#> v57         1        1        .        1        1
#> v58         1        1        .        1        1
#> v59         1        1        .        1        1
#> v60         1        1        .        1        1
#> v61         1        1        .        1        1
#> v62         1        1        .        1        1
#> v63         1        1        .        .        1
#> v64         1        1        .        .        1
#> v67         1        1        .        .        1
#> v68         1        1        .        1        1
#> v69         1        1        .        1        1
#> v70         1        1        .        1        1
#> v71         1        1        .        1        1
#> v72         1        1        .        1        1
#> v73         1        1        .        1        1
#> v74         1        1        .        1        1
#> v75         1        1        .        1        1
#> v76         1        1        .        .        1
#> v79         1        .        .        .        .
#> v80         1        1        .        1        1
#> v81         1        1        .        1        1
#> v82         1        1        .        1        1
#> v83         1        1        .        1        1
#> v84         1        1        .        1        1
#> v85         1        1        .        1        1
#> v86         1        1        .        1        1
#> v87         1        1        .        .        1
#> v88         1        1        .        .        1
#> v89         1        1        .        .        1
#> v92         1        .        .        .        .
#> v93         1        1        .        1        1
#> v94         1        1        .        1        1
#> v95         1        1        .        1        1
#> v96         1        1        .        1        1
#> v97         1        1        .        1        1
#> v98         1        1        .        .        1
#> v99         1        1        .        .        1
#> 
#> $poly_shp
#> class       : SpatialPolygonsDataFrame 
#> features    : 92 
#> extent      : 2.065453, 15.06545, 3.611903, 14.6119  (xmin, xmax, ymin, ymax)
#> crs         : NA 
#> variables   : 2
#> names       : grids, richness 
#> min values  :  v100,        1 
#> max values  :   v99,        5 
#> 

# }

s <- readRDS(system.file("nigeria/nigeria.rds", package = "phyloregion"))

set.seed(1)
m <- data.frame(sp::spsample(s, 10000, type = "nonaligned"))
names(m) <- c("lon", "lat")
species <- paste0("sp", sample(1:1000))
m$taxon <- sample(species, size = nrow(m), replace = TRUE)

pt <- points2comm(dat = m, mask = s, res = 0.5, lon = "lon", lat = "lat",
            species = "taxon")
head(pt[[1]])

#> 6 x 1000 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 1000 column names ‘sp1’, ‘sp10’, ‘sp100’ ... ]]

#>                                                                               
#> v100 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v103 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . 1 .
#> v101 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . 1 . 1 . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . 1 . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v103 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v105 . . . . . . 1 . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . 1 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . 1 . 1 . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . . . . 1 . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . 1 . . . . . . 1 . . . . . 1 . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . 1 1 . . 1 . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . 1 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 .
#> v102 . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . . . . . . . . .
#> v103 . . . . 1 . . . . . . . . . . . . . . . . 1 . . . 1 . . . . 1 . . 1 . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . 1 .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . 2 . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . 1 . . . . . . . . . . . . . . 1 . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . 1 . . 1 . . 1 . . 1 . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . 1
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . 1 . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v105 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . 1 .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . 1 . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . 2 . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . 1 . . . . . . 1 . 1 . . . . . . . . . . . . .
#>                                                                               
#> v100 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v103 . . . . . 1 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . 1 . . . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . 1 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . 1 . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . 1 . . 1 . . . . . . . . . . . . . . . .
#> v103 . . . . 1 1 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . 1 . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . 1 . . . . . 1 . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v103 . . . . . . . . . . 2 . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . 1 . . . . . . . . . . . . . . . . . 1 1 . . . . . . . 1 .
#> v101 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . .
#> v103 . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . 1 . . . . 1 . . . . . . . . . . 1 . . . . 1 . . . . . . 1 . . 1 .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . .
#> v103 . . . . . 1 . . . . 1 . . . . . . . 1 1 . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 1 . . . . . . . . . 1 . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . 1 . . . . . . . . 1 . . . . . . 1 . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#> v101 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#> v105 . . . 1 1 . . 1 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#> v103 . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v105 . . . . . . . . . . . . . . . . . . . . . 1 1 . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . 1 . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . 1 . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . 1 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . 1 1 . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . 1 . . . . . . . . 1 . . . 1 1 . . . 1 . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . 1 . . . 1 . . 1 . . . . . . . . . . 2 . . . . . . . 1 . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>       
#> v100 .
#> v101 .
#> v102 .
#> v103 .
#> v104 .
#> v105 .