Species distribution models for a range of algorithms

This function computes species distribution models using four modelling algorithms: generalized linear models, generalized boosted models, random forests, and maximum entropy (only if rJava is available). Note: this is an experimental function, and may change in the future.

sdm(
  x,
  pol = NULL,
  predictors = NULL,
  blank = NULL,
  tc = 2,
  lr = 0.001,
  bf = 0.75,
  n.trees = 50,
  step.size = n.trees,
  k = 5,
  herbarium.rm = FALSE,
  n.points = 80
)

Arguments

x
A dataframe containing the species occurrences and geographic coordinates. Column 1 labeled as "species", column 2 "lon", column 3 "lat".

pol
A polygon shapefile specifying the boundary to restrict the prediction. If not specified, a minimum convex polygon is estimated using the input data frame of species occurrences.

predictors
RasterStack of environmental descriptors on which the models will be projected

blank
A blank raster upon which the prediction layer is aggregated to.

tc
Integer. Tree complexity. Sets the complexity of individual trees

lr
Learning rate. Sets the weight applied to individual trees

bf
Bag fraction. Sets the proportion of observations used in selecting variables

n.trees
Number of initial trees to fit. Set at 50 by default

step.size
Number of trees to add at each cycle

k
Number of groups

herbarium.rm
Logical, remove points within 50 km of herbaria.

n.points
Minimum number of points required to successfully run a species distribution model

Value

A list with the following objects:

ensemble_raster The ensembled raster that predicts the potential species distribution.

ensemble_AUC The median AUCs of models.

data The dataframe that was used to implement the model.

indiv_models Raster layers for the separate models that predict the potential species distribution.

single_AUCs The AUCs for the seperate models.

References

Zurell, D., Franklin, J., König, C., Bouchet, P.J., Dormann, C.F., Elith, J., Fandos, G., Feng, X., Guillera‐Arroita, G., Guisan, A., Lahoz‐Monfort, J.J., Leitão, P.J., Park, D.S., Peterson, A.T., Rapacciuolo, G., Schmatz, D.R., Schröder, B., Serra‐Diaz, J.M., Thuiller, W., Yates, K.L., Zimmermann, N.E. and Merow, C. (2020), A standard protocol for reporting species distribution models. Ecography, 43: 1261-1277.

Examples

# \donttest{
library(raster)
# get predictor variables
f <- list.files(path=paste(system.file(package="phyloregion"),
                '/ex', sep=''), pattern='.tif', full.names=TRUE)

preds <- stack(f)
#plot(preds)
# get species occurrences
d <- read.csv(system.file("ex/Bombax.csv", package="phyloregion"))

# fit ensemble model for four algorithms
mod <- sdm(d, predictors = preds)

#> Loading required namespace: rJava

#> This is MaxEnt version 3.4.1 

#> Warning: 1 (1.12%) of the presence points have NA predictor values

#> Warning: only got:197random background point values; Small exent? Or is there a layer with many NA values?

#> Loading required namespace: gbm

#> Error in dismo::gbm.step(data = envtrain, gbm.x = Preds, gbm.y = y, family = "bernoulli",     tree.complexity = tc, learning.rate = lr, bag.fraction = bf,     verbose = TRUE, silent = TRUE, plot.main = FALSE, step.size = step.size,     n.trees = n.trees): you need to install the gbm package to run this function

# }