Species' evolutionary distinctiveness

Calculates evolutionary distinctiveness measures for a suite of species by: a) equal splits (Redding and Mooers 2006) b) fair proportions (Isaac et al., 2007). This a new implementation of the picante function evol.distinct however allowing multifurcations and can be orders of magnitude faster.

evol_distinct(
  tree,
  type = c("equal.splits", "fair.proportion"),
  scale = FALSE,
  use.branch.lengths = TRUE,
  ...
)

Arguments

tree

an object of class phylo.

type

a) equal splits (Redding and Mooers 2006) or b) fair proportions (Isaac et al., 2007)

scale

The scale option refers to whether or not the phylogeny should be scaled to a depth of 1 or, in the case of an ultrametric tree, scaled such that branch lengths are relative.

use.branch.lengths

If use.branch.lengths=FALSE, then all branch lengths are changed to 1.

...

Further arguments passed to or from other methods.

Value

a named vector with species scores.

References

Redding, D.W. and Mooers, A.O. (2006). Incorporating evolutionary measures into conservation prioritisation. Conservation Biology, 20, 1670--1678.

Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C. and Baillie, J.E.M. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE, 2, e296.

See also

Author

Klaus Schliep

Examples

tree <- ape::rcoal(10)
evol_distinct(tree)
#>        t2        t7       t10        t1        t9        t6        t8        t5 
#> 2.2643842 1.4752927 0.8398421 0.4971818 0.4971818 1.5015497 0.4257105 0.4257105 
#>        t4        t3 
#> 0.3955691 0.3955691 
evol_distinct(tree, type = "fair.proportion")
#>        t2        t7       t10        t1        t9        t6        t8        t5 
#> 1.9474928 1.1599697 0.8387633 0.8138283 0.8138283 0.7269576 0.6193585 0.6193585 
#>        t4        t3 
#> 0.5892171 0.5892171