Create a fishnet of regular grids

The fishnet function creates a regular grid of locations covering the study area at various grain sizes.

fishnet(mask, res = 0.5, type = "square")

Arguments

mask
a polygon shapefile covering the boundary of the survey region.

res
the grain size of the grid cells in decimal degrees (default).

type
the type of grid cell; either “square” (default) for square grids, or “hexagon” for hexagonal grids.

Value

A spatial polygon object of equal area grid cells covering the defined area.

References

Phillips, S.J., Anderson, R.P. & Schapire, R.E. (2006) Maximum entropy modeling of species geographic distributions. Ecological Modelling 190: 231-259.

Examples

file <- system.file("nigeria/nigeria.rds", package="phyloregion")
d <- readRDS(file)
d1 <- fishnet(d, res = 0.75)