Convert raw input distribution data to community

The functions points2comm, polys2comm, raster2comm provide convenient interfaces to convert raw distribution data often available as point records, extent-of-occurrence polygons and raster layers, respectively, to a community composition data frame at varying spatial grains and extents for downstream analyses.

raster2comm(files)

polys2comm(dat, res = 1, shp.grids = NULL, species = "species", ...)

points2comm(
  dat,
  mask = NULL,
  res = 1,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  shp.grids = NULL,
  index = "taxon_richness",
  ...
)

Arguments

files
list of raster layer objects with the same spatial extent and resolution.

dat
layers of merged maps corresponding to species ranges for polys2comm; or point occurrence data frame for points2comm, with at least three columns:

Column 1: species (listing the taxon names)

Column 2: decimallongitude (corresponding to decimal longitude)

Column 3: decimallatitude (corresponding to decimal latitude)

res
the grain size of the grid cells in decimal degrees (default).

shp.grids
if specified, the polygon shapefile of grid cells with a column labeled “grids”.

species
a character string. The column with the species or taxon name. Default = “species”.

...
Further arguments passed to or from other methods.

mask
Only applicable to points2comm. If supplied, a polygon shapefile covering the boundary of the survey region.

lon
character with the column name of the longitude.

lat
character with the column name of the latitude.

index
Diversity index, one of “richness” (default) or “abundance”.

Value

comm_dat: community data frame

poly_shp: shapefile of grid cells with the values per cell.

Examples

if (FALSE) {
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
raster2comm(files)
}

if (FALSE) {
s <- readRDS(system.file("nigeria/nigeria.rds", package="phyloregion"))
sp <- random_species(100, species=5, shp=s)
polys2comm(dat = sp, species = "species")
}

s <- readRDS(system.file("nigeria/nigeria.rds", package = "phyloregion"))

set.seed(1)
m <- data.frame(sp::spsample(s, 10000, type = "nonaligned"))
names(m) <- c("lon", "lat")
species <- paste0("sp", sample(1:1000))
m$taxon <- sample(species, size = nrow(m), replace = TRUE)

points2comm(dat = m, mask = s, res = 0.5, lon = "lon", lat = "lat",
            species = "taxon")

#> $comm_dat
#>       grids species
#>    1:  v439   sp118
#>    2:  v439   sp826
#>    3:  v439   sp121
#>    4:  v440    sp58
#>    5:  v440   sp585
#>   ---              
#> 9805:    v6   sp413
#> 9806:    v7   sp322
#> 9807:    v6   sp970
#> 9808:    v6   sp106
#> 9809:    v7   sp856
#> 
#> $poly_shp
#> class       : SpatialPolygonsDataFrame 
#> features    : 332 
#> extent      : 2.66843, 14.66843, 4.39201, 13.89201  (xmin, xmax, ymin, ymax)
#> crs         : NA 
#> variables   : 2
#> names       : grids, richness 
#> min values  :  v100,        1 
#> max values  :   v99,       36 
#>