Convert raw input distribution data to community

The functions points2comm, polys2comm, raster2comm provide convenient interfaces to convert raw distribution data often available as point records, polygons and raster layers, respectively, to a community composition data frame at varying spatial grains and extents for downstream analyses.

raster2comm(files)

polys2comm(dat, res = 1, species = "species", shp.grids = NULL, ...)

points2comm(
  dat,
  mask = NULL,
  res = 1,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  shp.grids = NULL,
  ...
)

Arguments

files
list of raster layer objects with the same spatial extent and resolution.

dat
layers of merged maps corresponding to species polygons for polys2comm; or point occurrence data frame for points2comm, with at least three columns:

Column 1: species (listing the taxon names)

Column 2: decimallongitude (corresponding to decimal longitude)

Column 3: decimallatitude (corresponding to decimal latitude)

res
the grain size of the grid cells in decimal degrees (default).

species
a character string. The column with the species or taxon name. Default = “species”.

shp.grids
if specified, the polygon shapefile of grid cells with a column labeled “grids”.

...
Further arguments passed to or from other methods.

mask
Only applicable to points2comm. If supplied, a polygon shapefile covering the boundary of the survey region.

lon
character with the column name of the longitude.

lat
character with the column name of the latitude.

Value

Each of these functions generate a list of two objects as follows:

comm_dat: (sparse) community matrix

poly_shp: shapefile of grid cells with the values per cell.

See also

Examples

# \donttest{
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
ras <- raster2comm(files) # Note, this function generates

#>   |                                                |                                        |   0%  |                                                |=====                                   |  12%  |                                                |==========                              |  25%  |                                                |===============                         |  38%  |                                                |====================                    |  50%  |                                                |=========================               |  62%  |                                                |==============================          |  75%  |                                                |===================================     |  88%  |                                                |========================================| 100%

     # a list of two objects
head(ras[[1]])

#> 6 x 8 sparse Matrix of class "dgCMatrix"
#>      Chytranthus_gilletii Commelina_ramulosa Cymbopogon_caesius
#> v100                    .                  .                  .
#> v101                    .                  .                  .
#> v102                    .                  .                  .
#> v103                    .                  .                  .
#> v104                    .                  .                  .
#> v105                    .                  .                  .
#>      Dalechampia_ipomoeifolia Grewia_barombiensis Indigofera_berhautiana
#> v100                        1                   .                      .
#> v101                        1                   .                      .
#> v102                        1                   .                      .
#> v103                        1                   .                      .
#> v104                        1                   .                      .
#> v105                        1                   .                      .
#>      Indigofera_erecta Indigofera_oubanguiensis
#> v100                 .                        .
#> v101                 .                        .
#> v102                 .                        .
#> v103                 .                        .
#> v104                 .                        .
#> v105                 .                        .

# }

# \donttest{
s <- readRDS(system.file("nigeria/nigeria.rds", package="phyloregion"))
sp <- random_species(100, species=5, shp=s)
pol <- polys2comm(dat = sp, species = "species")

#>   |                                                |                                        |   0%  |                                                |========                                |  20%  |                                                |================                        |  40%  |                                                |========================                |  60%  |                                                |================================        |  80%  |                                                |========================================| 100%

head(pol[[1]])

#> 6 x 5 sparse Matrix of class "dgCMatrix"
#>        species1 species2 species3 species4 species5
#> v27905        1        .        .        1        1
#> v27906        1        1        1        1        1
#> v27907        1        1        1        1        1
#> v27908        1        1        1        1        1
#> v27909        1        1        1        1        1
#> v27910        1        1        1        1        1

# }

s <- readRDS(system.file("nigeria/nigeria.rds", package = "phyloregion"))

set.seed(1)
m <- data.frame(sp::spsample(s, 10000, type = "nonaligned"))
names(m) <- c("lon", "lat")
species <- paste0("sp", sample(1:1000))
m$taxon <- sample(species, size = nrow(m), replace = TRUE)

pt <- points2comm(dat = m, mask = s, res = 0.5, lon = "lon", lat = "lat",
            species = "taxon") # Note, this generates a list of two objects
head(pt[[1]])

#> 6 x 1000 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 1000 column names ‘sp1’, ‘sp10’, ‘sp100’ ... ]]

#>                                                                               
#> v100 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v103 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . 1 .
#> v101 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . 1 . 1 . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . 1 . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v103 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v105 . . . . . . 1 . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . 1 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . 1 . 1 . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . . . . 1 . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . 1 . . . . . . 1 . . . . . 1 . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . 1 1 . . 1 . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . 1 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 .
#> v102 . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . . . . . . . . .
#> v103 . . . . 1 . . . . . . . . . . . . . . . . 1 . . . 1 . . . . 1 . . 1 . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . 1 .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . 2 . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . 1 . . . . . . . . . . . . . . 1 . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . 1 . . 1 . . 1 . . 1 . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . 1
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . 1 . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v105 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . 1 .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . 1 . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . 2 . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . 1 . . . . . . 1 . 1 . . . . . . . . . . . . .
#>                                                                               
#> v100 . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v103 . . . . . 1 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . 1 . . . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . 1 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . 1 . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . 1 . . 1 . . . . . . . . . . . . . . . .
#> v103 . . . . 1 1 . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . 1 . . . . 1 . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . 1 . . . . . 1 . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . . . . . . .
#> v103 . . . . . . . . . . 2 . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . 1 . . . . . . . . . . . . . . . . . 1 1 . . . . . . . 1 .
#> v101 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . .
#> v103 . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . 1 . . . . 1 . . . . . . . . . . 1 . . . . 1 . . . . . . 1 . . 1 .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . .
#> v103 . . . . . 1 . . . . 1 . . . . . . . 1 1 . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 1 . . . . . . . . . 1 . . . . . . . . . . . . 1 . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . 1 . . . . . . . . 1 . . . . . . 1 . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#>                                                                               
#> v100 . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#> v101 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#> v105 . . . 1 1 . . 1 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#> v103 . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . .
#> v105 . . . . . . . . . . . . . . . . . . . . . 1 1 . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . 1 . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . 1 . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1
#>                                                                               
#> v100 . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> v101 . . 1 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . 1 1 . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v101 . . . . . . . . 1 . . . . . . . . 1 . . . 1 1 . . . 1 . . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> v100 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v101 . 1 . . . 1 . . 1 . . . . . . . . . . 2 . . . . . . . 1 . . . . . . . . .
#> v102 . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> v103 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> v104 . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . .
#> v105 . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#>       
#> v100 .
#> v101 .
#> v102 .
#> v103 .
#> v104 .
#> v105 .